third_party/blink/web_tests/external/wpt/css/selectors/parsing/parse-has.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] ":has(:is(.a, 123))" should be a valid selector
  assert_equals: serialization should be canonical expected ":has(:is(.a, 123))" but got ":has(:is(.a))"
Harness: the test ran to completion.
